latesttry/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==2.2.2
gunicorn==20.1.0
Flask
tensorflow=2.12.0
keras
Pillow
